Add files via upload
Immigrants, refugees and refugee claimants make 40 percent of people using shelters in the city of Toronto that also identify as homeless. (Teresa, 2019). With the public’s growing concerns on why the government of Canada is delaying to declare homelessness as a state of emergency (PressProgress, 2019), one of the homeless shelter’s greatest threats is capacity. Using the daily shelter occupancy dataset from the city of Toronto’s Open Data portal, a study of the dataset will be provided to explore the Question of Which shelter sector (male/female/Co-ed/families/youth) has the highest occupancy?